requirements.txt
boto3
botocore
cryptography
deepface
numpy
opencv_python
pandas
Pillow
PyQt5
PyQt5_sip
SQLAlchemy
tqdm
ruff
